[RHTAP-4916] Moves componentActionStrategy under common strategies
