+++ src/rhtap/common/strategies/componentActionStrategy.ts
