+++ src/rhtap/postcreation/strategies/gitlabCIPostCreateActionStrategy.ts
@@ -3,7 +3,7 @@ import { GitType } from '../../core/integration/git';
 import { AddGitlabProjectVariablesCommand } from './commands/addGitlabProjectVariablesCommand';
 import { Command } from './commands/command';
 import { UpdateCIRunnerImage } from './commands/updateCIRunnerImage';
-import { ComponentActionStrategy } from './componentActionStrategy';
+import { ComponentActionStrategy } from '../../common/strategies/componentActionStrategy';
 
 /**
  * GitLab-specific implementation of post-creation action strategy
